Allow concurrency: lock sandbox when cleaning up
This change, if applied, allows multiple instances of conan to run
in parallel.

This is done by locking the sandbox that is being cleaned up.

- Set column conan_status to 'cleanup in progress'
- Set conan_timestamp to now()
- Set conan_hostname to hostname, it can be useful for troublehooting
  to know on what host it's currently running
- A lock timeouts after N hours. Set the default to 2h.

Define new parameter: lock_timeout

Lock timeout:  the number of hours after which a lock on a sandbox expires.
For ex: '2': a conan process will have 2h to cleanup the sandbox before another
process can claim the sandbox for cleanup.
That parameter prevent a sandbox from being locked forever if something goes wrong with
the conan process owning the lock.

Also change all the date format to iso-8601.