conan: Enable separate dynamoDB for DEV
To be able to cleanup the dev database, we need to pass the AWS profile
to use for any dynamodb communication.

This change, if appliced, introduce a new environment variable 'dynamodb_profile'
to be able to specify the AWS profile to use.

By default, it's the same as the AWS profile to manage all the accounts
and to assume role inside the accounts (prod).

For DEV, a different AWS account will be passed, as the DB is in a
different AWS account.

Also add a variable for the name of the instance.

Update the conan-dev.service systemd service unit.