Fix: Ignore accounts marked for cleanup
Re-using the same uuid in functional tests releaved a minor bug where
listings of resources or jobs of a placement (with same UUID) are polluted by old
accounts still marked for cleanup.

This commit introduces new functions `FetchAllActiveByUuid` to ignore
accounts already marked for cleanup and use that
instead of the previous functions.