Fix Sandbox API scheduling of OcpSandbox
see GPTEINFRA-9992

The current scheduling algorithm of OcpSandbox selects the cluster with the least memory usage.
Plus, the calculation is done by summing pods requests memory, which seems dubious.

This is counterintuitive and against the principle of least astonishment.

Instead, the algorithm should randomly select a cluster with available memory space. The minimum memory space required for a cluster to be schedulable should be configurable in the OcpSharedClusterConfiguration table. The default value should be 90% of the total memory of the cluster.

This change:

* Create 2 new columns in the `ocp_shared_cluster_configuration` for max
CPU and max Memory percentage for a cluster to be considered healthy.
  max CPU default: 100
  max Memory default: 90
* Upgrade to Go 1.22, required by k8s.io module `k8s.io/api@v0.30.2`
* Schedule randomly on the first available cluster.
* TODO: loop only on schedulable nodes