Add bool SkipQuota to OcpSharedClusterConf
SkipQuota disables quota creation entirely. That will be useful to push the new code to production without any change to current behavior. This way we can set this value on all existing clusters to false as a first step, and enable it as we go. Later we can turn this feature on by default.