feat: Add console_url to OcpSandbox + auto-detect (#152)
This pull request introduces the `console_url` field to the OcpSandbox resource.

This change is necessary because the OpenShift console cannot always be forged downstream, and may be exposed on a different domain than the cluster's primary applications ingress domain. By explicitly providing the console_url, we ensure that users can reliably access the web console for their sandbox environment.

The implementation first checks if the `console_url` is defined in the `additionalVars` of the cluster configuration. If it's not present, it automatically detects the URL by querying the Kubernetes API for the console route within the openshift-console namespace.

This change also includes:

* Updates to the Swagger API documentation
* New assertions in the Hurl tests to validate the presence and format of the `console_url` in the API responses.
  * Test the console URL retruned, make sure it's returning a 200 OK